+++ R/update_presentation_state.R
@@ -16,12 +16,12 @@ update_presentation_state <- function(presentation_id) {
 
   # Print to console anything that has changed
 
-  # All of these depend on a GET call to query2
+  # All of these depend on a GET call to query
 }
 
 
 
 has_updates <- function(presentation_id) {
   # Check that the revision_id is the same
-  # All of these depend on a GET call to query2
+  # All of these depend on a GET call to query
 }